build: addresses comments refactors logic
