+++ packages/features/portfolio/src/hooks/index.ts
@@ -4,3 +4,4 @@ export * from './use-txn-history-table';
 export * from './use-portfolio-side-drawers';
 export * from './use-txn-history-button-bar';
 export * from './use-txn-history-dropdowns';
+export * from './use-txn-history-data';
