+++ packages/features/portfolio/src/hooks/use-txn-history-table.tsx
@@ -9,27 +9,15 @@ import {
 } from '@notional-finance/mui';
 import { Box, useTheme } from '@mui/material';
 import { TXN_HISTORY_TYPE } from '@notional-finance/shared-config';
-import {
-  getEtherscanLink,
-  formatTokenType,
-  formatTokenAmount,
-  getDateString,
-  formatNumber,
-} from '@notional-finance/helpers';
-import {
-  useSelectedNetwork,
-  useTransactionHistory,
-} from '@notional-finance/notionable-hooks';
+import { getDateString } from '@notional-finance/helpers';
 import { FormattedMessage } from 'react-intl';
 
 export const useTxnHistoryTable = (
   currencyOptions: string[],
   assetOrVaultOptions: string[],
-  txnHistoryType: TXN_HISTORY_TYPE
+  txnHistoryType: TXN_HISTORY_TYPE,
+  accountHistoryData: any[]
 ) => {
-  const accountHistory = useTransactionHistory();
-  const selectedNetwork = useSelectedNetwork();
-
   const VaultNameCell = ({ cell: { value } }) => {
     const theme = useTheme();
     return (
@@ -122,65 +110,6 @@ export const useTxnHistoryTable = (
     },
   ];
 
-  const allAccountHistoryData = accountHistory
-    .sort((x, y) => y.timestamp - x.timestamp)
-    .map(
-      ({
-        bundleName,
-        underlyingAmountRealized,
-        tokenAmount,
-        token,
-        realizedPrice,
-        timestamp,
-        transactionHash,
-        underlying,
-        impliedFixedRate,
-        vaultName,
-      }) => {
-        const assetData = formatTokenType(token);
-        return {
-          transactionType: {
-            label: bundleName,
-            symbol: underlying.symbol.toLowerCase(),
-          },
-          vaultName: vaultName,
-          underlyingAmount: underlyingAmountRealized.toDisplayStringWithSymbol(
-            3,
-            true
-          ),
-          assetAmount: formatTokenAmount(tokenAmount, impliedFixedRate),
-          asset: {
-            label: assetData.title,
-            symbol: assetData.title.toLowerCase(),
-            caption: assetData.caption ? assetData.caption : '',
-          },
-          price: formatNumber(realizedPrice.toFloat(), 2),
-          time: timestamp,
-          txLink: {
-            hash: transactionHash,
-            href: getEtherscanLink(transactionHash, selectedNetwork),
-          },
-          currency: underlying.symbol,
-        };
-      }
-    );
-
-  const formatTxnHistoryData = () => {
-    const porfolioHoldingsData = allAccountHistoryData.filter(
-      ({ vaultName }) => !vaultName
-    );
-
-    const leveragedVaultsData = allAccountHistoryData.filter(
-      ({ vaultName }) => vaultName
-    );
-
-    return txnHistoryType === TXN_HISTORY_TYPE.PORTFOLIO_HOLDINGS
-      ? porfolioHoldingsData
-      : leveragedVaultsData;
-  };
-
-  const initialData = formatTxnHistoryData();
-
   const txnHistoryColumns =
     txnHistoryType === TXN_HISTORY_TYPE.PORTFOLIO_HOLDINGS
       ? tableColumns.filter(({ accessor }) => accessor !== 'vaultName')
@@ -188,29 +117,28 @@ export const useTxnHistoryTable = (
 
   const filterTxnHistoryData = () => {
     const filterData = [...currencyOptions, ...assetOrVaultOptions];
-    if (filterData.length === 0) return initialData;
+    if (filterData.length === 0) return accountHistoryData;
     if (assetOrVaultOptions.length > 0 && currencyOptions.length > 0) {
       return txnHistoryType === TXN_HISTORY_TYPE.PORTFOLIO_HOLDINGS
-        ? initialData
+        ? accountHistoryData
             .filter(({ currency }) => filterData.includes(currency))
             .filter(({ asset }) => filterData.includes(asset?.label))
-        : initialData
+        : accountHistoryData
             .filter(({ currency }) => filterData.includes(currency))
             .filter(
               ({ vaultName }) => vaultName && filterData.includes(vaultName)
             );
     }
     if (currencyOptions.length > 0) {
-      return initialData.filter(({ currency }) =>
+      return accountHistoryData.filter(({ currency }) =>
         currencyOptions.includes(currency)
       );
     }
-
     if (
       assetOrVaultOptions.length > 0 &&
       txnHistoryType === TXN_HISTORY_TYPE.PORTFOLIO_HOLDINGS
     ) {
-      return initialData.filter(({ asset }) =>
+      return accountHistoryData.filter(({ asset }) =>
         filterData.includes(asset?.label)
       );
     }
@@ -218,13 +146,11 @@ export const useTxnHistoryTable = (
       assetOrVaultOptions.length > 0 &&
       txnHistoryType === TXN_HISTORY_TYPE.LEVERAGED_VAULT
     ) {
-      return initialData.filter(
+      return accountHistoryData.filter(
         ({ vaultName }) => vaultName && filterData.includes(vaultName)
       );
     }
   };
-
-  const txnHistoryData = filterTxnHistoryData();
 
   const marketDataCSVFormatter = useCallback((data: any[]) => {
     return data.map(
@@ -255,6 +181,8 @@ export const useTxnHistoryTable = (
       }
     );
   }, []);
+
+  const txnHistoryData = filterTxnHistoryData();
 
   return { txnHistoryData, txnHistoryColumns, marketDataCSVFormatter };
 };
